Implement preservable jobs and configurable jobs table name (#11)
Preserved jobs, if enabled, are marked with a `completed` state in the
database. Turning preservation off if it was previously on does _not_
clear old jobs from the table; it just prevents new ones from being
preserved.

Configuring the jobs table name requires passing the same name (and
optionally, space) to both the queue driver _and_ the migration. This is
necessary because there's no way for the migration to see the driver's
configuration.

Closes #9.
Closes #10.