Fix SQLite with delayed jobs (#16)
Currently, jobs with a value for `delay_until` are never run when using
SQLite as the underlying database. This is because one of the conditions
used to select a job is `unixepoch(coalesce("delay_until", 'now')) <=
unixepoch('now')`. This is fine if `delay_until` is NULL, since it will
be equivalent to `unixepoch('now') <= unixepoch('now')` which is always
true. However, if `delay_until` has a value, the condition becomes
equivalent to `unixepoch("delay_until") <= unixepoch('now')` which is
incorrect since `delay_until` is already stored as an epoch timestamp.
Passing a epoch timestamp to `unixepoch` results in NULL.