Fix jobs being retried forever (#15)
Currently, if a job is dispatched with a non-zero maxRetryCount, the job
is retried forever. This is because the `JobData` was initialised with
the default value (0) for `attempts`.

I’ve added a test for this, however it is currently flakey (especially
when run with SQLite) since `delayUntil` is set for retried jobs, even
if the retry delay is zero. This can be fixed by changing the behaviour
in the queues package to set a nil `delayUntil` when the retry delay is
zero, but I’m not sure if this is the best solution.

---------

Co-authored-by: Gwynne Raskind <gwynne@vapor.codes>