Fix creation of the database enum when using a non-default space (#17)
Previously, if a non-`nil` `jobsTableSpace` was provided to the
`JobModelMigration` when using PostgreSQL, the job state enum type was
incorrectly created in the default schema instead of alongside the jobs
table. This is now fixed, and the name of the space is no longer used to
prefix the names of the enumeration or the index on the table. The
migration's `revert(_:)` method will attempt to drop the old name from
the default space (using `IF EXISTS`) in order to maintain
compatibility.

Additional changes:

- `Package.swift` no longer imports `Foundation` to get at the current
environment (I learned that `Context.environment` exists).
- The minimum version requirements for the dependencies have been
bumped.
- The CI has been updated to include tests against the latest PostgreSQL
and MySQL versions.
- A number of issues in the README have been fixed and information on
the driver's configurable options has been added.