Add migration to support users of old versions (#18)
There is now a new migration, `JobModelOldFormatMigration`, which can be
used in place of `JobModelMigration` by users who have an existing job
metadata table from a 1.x or 2.x version of the driver. This migration
will upgrade old tables to the new format without any loss of data. See
the README and the comments in the code for more details.

Additional changes:
- Swift 5.10 is now the required minimum version.
- The code is now compliant with `MemberImportVisibility`.
- For MySQL users, the main migration will now use `DATETIME(6)` instead
of `DATETIME` so that timestamps have subsecond precision. The
`updated_at` column is now also a `DATETIME(6)` instead of a
`TIMESTAMP`. Existing users do not need to make any changes to their
tables or add any new migrations; the existing table layout continues to
work as it did before. Only new users and users who change their tables
manually will see this.
- For non-MySQL users, the `updated_at` column is now `NOT NULL`. Again,
existing users do not need to take any action; everything continues to
work as before.
- The CI has been updated for Swift 6.1 and now checks Linux SDK
compatibility.
- A lingering `Sendable` warning has been fixed.